refactor component name and directory structure
