+++ src/components/detail/PostDetail.tsx
@@ -1,24 +1,20 @@
 "use client";
 
-/**
- * This code was generated by v0 by Vercel.
- * @see https://v0.dev/t/wsZsVmQBbbr
- */
 import { AvatarImage, AvatarFallback, Avatar } from "@/components/ui/avatar";
 import { Input } from "@/components/ui/input";
 import ReactCardFlip from "react-card-flip";
 import { useState } from "react";
-import DetailFront from "./DetailFront";
-import DetailBack from "./DetailBack";
 import { OnePost } from "@/lib/definitions";
 import { useRouter } from "next/navigation";
+import DetailImageFront from "./DetailImageFront";
+import DetailImageBack from "./DetaiImagelBack";
 
 type Props = {
   post: OnePost;
   myId: string | undefined;
 };
 
-export function PostModal(props: Props) {
+export function PostDetail(props: Props) {
   const { post, myId } = props;
   const [isFlipped, setIsFlipped] = useState(false);
 
@@ -36,8 +32,8 @@ export function PostModal(props: Props) {
           flipSpeedFrontToBack={0.6}
           infinite={true}
         >
-          <DetailFront src={post.imgFront!} handleClick={handleClick} />
-          <DetailBack
+          <DetailImageFront src={post.imgFront!} handleClick={handleClick} />
+          <DetailImageBack
             src={post.imgBack!}
             myId={myId}
             userId={post.authorId!}
